requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
sentence-transformers
faiss-cpu
yfinance 
pandas 
newsapi-python 
textblob
google-genai